refactor(development-harness): split GitHubBackend into composed collaborators
GitHubBackend combined issue snapshot/reconciliation, content migration
orchestration, legacy content persistence, and provider composition in one
1549-line class. Extract each responsibility into its own collaborator that
GitHubBackend composes, following the _GitHubContentsStore precedent.

- github_content_stores.py: the legacy Gist-backed plan and dispatch stores
  plus the shared content-store protocol, revision, owner-reference, and
  full-page enumeration helpers. Removes the back-reference these stores had
  to GitHubBackend._content_revision.
- github_content_migration.py: _GitHubContentMigration resolves reads, writes,
  and discovery across the Contents API and the legacy stores and owns the
  migration compare-and-swap; _GitHubContentCache owns the offline, replay,
  and staleness policy wrapped around it.
- github_work_items.py: _GitHubWorkItemSync translates issues into provider
  snapshots and applies provider patches; _GitHubReconciliation drives the
  reconciliation cycle against the provider-private FileCache.

GitHubBackend keeps the GitHub API delegation surface (gh_client, github_sync,
github_branches, rendering, and the bounded aliased targeted-issue query) and
becomes the composition root: every protocol method delegates to the
collaborator that owns it. The collaborators reach GitHub and the injected
stores back through narrow protocols GitHubBackend satisfies structurally, so
each gateway call and injected store stays substitutable on the backend
instance rather than captured at construction.

Public protocol behavior is unchanged; no test was modified.
